+++ scripts/libwrapper.js
@@ -2,7 +2,7 @@ import CONSTANTS from "./constants.js";
 import { libWrapper } from "./lib/libWrapper/shim.js";
 
 export default function registerLibwrappers(){
-  libWrapper.register(CONSTANTS.MODULE_NAME, "PIXI.resources.BaseImageResource.prototype.upload", PIXIUPLOAD);
+  libWrapper.register(CONSTANTS.MODULE_NAME, "PIXI.BaseImageResource.prototype.upload", PIXIUPLOAD);
 }
 
 function PIXIUPLOAD(wrapped, ...args){