+++ docs/changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### Version 2.412
+- *Animation* - Fixed `.moveTowards()` going into infinite loop if the source and targets are on top of each other
 - *Effects* - Fixed `.shape()` taking grid size into account multiple times
 
 ### Version 2.411